Crate storage1

Features

Handle the following Resources with ease from the central hub ...

bucket access controls (delete, get, insert, list, patch and update)

buckets (delete, get, insert, list, patch and update)

channels (stop)

default object access controls (delete, get, insert, list, patch and update)

object access controls (delete, get, insert, list, patch and update)

objects (compose, copy, delete, get, insert, list, patch, update and watch all)

Everything else about the storage API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.object_access_controls().get(...).doit()
let r = hub.default_object_access_controls().patch(...).doit()
let r = hub.object_access_controls().update(...).doit()
let r = hub.object_access_controls().list(...).doit()
let r = hub.object_access_controls().patch(...).doit()
let r = hub.default_object_access_controls().update(...).doit()
let r = hub.default_object_access_controls().insert(...).doit()
let r = hub.object_access_controls().insert(...).doit()
let r = hub.object_access_controls().delete(...).doit()
let r = hub.default_object_access_controls().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
storage1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate storage1;
use storage1::ObjectAccessControl;
use storage1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use storage1::Storage;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Storage::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: ObjectAccessControl = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.object_access_controls().update(&req, "bucket", "object", "entity")
             .generation("accusam")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Bucket
A bucket.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update buckets (request|response)

insert buckets (request|response)

delete buckets (none)

patch buckets (request|response)

list buckets (none)

get buckets (response)

BucketAccessControl
An access-control entry.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list bucket access controls (none)

patch bucket access controls (request|response)

delete bucket access controls (none)

update bucket access controls (request|response)

get bucket access controls (response)

insert bucket access controls (request|response)

BucketAccessControlDeleteMethodBuilder
Permanently deletes the ACL entry for the specified entity on the specified bucket.

BucketAccessControlGetMethodBuilder
Returns the ACL entry for the specified entity on the specified bucket.

BucketAccessControlInsertMethodBuilder
Creates a new ACL entry on the specified bucket.

BucketAccessControlListMethodBuilder
Retrieves ACL entries on the specified bucket.

BucketAccessControlMethodsBuilder
A builder providing access to all methods supported on bucketAccessControl resources. It is not used directly, but through the Storage hub.

BucketAccessControlPatchMethodBuilder
Updates an ACL entry on the specified bucket. This method supports patch semantics.

BucketAccessControlProjectTeam
The project team associated with the entity, if any.

This type is not used in any activity, and only used as part of another schema.

BucketAccessControlUpdateMethodBuilder
Updates an ACL entry on the specified bucket.

BucketAccessControls
An access-control list.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list bucket access controls (response)

BucketCors
The bucket's Cross-Origin Resource Sharing (CORS) configuration.

This type is not used in any activity, and only used as part of another schema.

BucketDeleteMethodBuilder
Permanently deletes an empty bucket.

BucketGetMethodBuilder
Returns metadata for the specified bucket.

BucketInsertMethodBuilder
Creates a new bucket.

BucketLifecycle
The bucket's lifecycle configuration. See lifecycle management for more information.

This type is not used in any activity, and only used as part of another schema.

BucketLifecycleRule
A lifecycle management rule, which is made of an action to take and the condition(s) under which the action will be taken.

This type is not used in any activity, and only used as part of another schema.

BucketLifecycleRuleAction
The action to take.

This type is not used in any activity, and only used as part of another schema.

BucketLifecycleRuleCondition
The condition(s) under which the action will be taken.

This type is not used in any activity, and only used as part of another schema.

BucketListMethodBuilder
Retrieves a list of buckets for a given project.

BucketLogging
The bucket's logging configuration, which defines the destination bucket and optional name prefix for the current bucket's logs.

This type is not used in any activity, and only used as part of another schema.

BucketMethodsBuilder
A builder providing access to all methods supported on bucket resources. It is not used directly, but through the Storage hub.

BucketOwner
The owner of the bucket. This is always the project team's owner group.

This type is not used in any activity, and only used as part of another schema.

BucketPatchMethodBuilder
Updates a bucket. This method supports patch semantics.

BucketUpdateMethodBuilder
Updates a bucket.

BucketVersioning
The bucket's versioning configuration.

This type is not used in any activity, and only used as part of another schema.

BucketWebsite
The bucket's website configuration.

This type is not used in any activity, and only used as part of another schema.

Buckets
A list of buckets.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list buckets (response)

Channel
An notification channel used to watch for resource changes.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

stop channels (request)

watch all objects (request|response)

ChannelMethodsBuilder
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Storage hub.

ChannelStopMethodBuilder
Stop watching resources through this channel

ComposeRequest
A Compose request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

compose objects (request)

ComposeRequestSourceObjects
The list of source objects that will be concatenated into a single object.

This type is not used in any activity, and only used as part of another schema.

ComposeRequestSourceObjectsObjectPreconditions
Conditions that must be met for this operation to execute.

This type is not used in any activity, and only used as part of another schema.

DefaultObjectAccessControlDeleteMethodBuilder
Permanently deletes the default object ACL entry for the specified entity on the specified bucket.

DefaultObjectAccessControlGetMethodBuilder
Returns the default object ACL entry for the specified entity on the specified bucket.

DefaultObjectAccessControlInsertMethodBuilder
Creates a new default object ACL entry on the specified bucket.

DefaultObjectAccessControlListMethodBuilder
Retrieves default object ACL entries on the specified bucket.

DefaultObjectAccessControlMethodsBuilder
A builder providing access to all methods supported on defaultObjectAccessControl resources. It is not used directly, but through the Storage hub.

DefaultObjectAccessControlPatchMethodBuilder
Updates a default object ACL entry on the specified bucket. This method supports patch semantics.

DefaultObjectAccessControlUpdateMethodBuilder
Updates a default object ACL entry on the specified bucket.

Object
An object.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list objects (none)

update objects (request|response)

copy objects (request|response)

watch all objects (none)

get objects (response)

insert objects (request|response)

compose objects (response)

delete objects (none)

patch objects (request|response)

ObjectAccessControl
An access-control entry.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get object access controls (response)

patch default object access controls (request|response)

update object access controls (request|response)

list object access controls (none)

patch object access controls (request|response)

update default object access controls (request|response)

insert default object access controls (request|response)

insert object access controls (request|response)

delete object access controls (none)

get default object access controls (response)

ObjectAccessControlDeleteMethodBuilder
Permanently deletes the ACL entry for the specified entity on the specified object.

ObjectAccessControlGetMethodBuilder
Returns the ACL entry for the specified entity on the specified object.

ObjectAccessControlInsertMethodBuilder
Creates a new ACL entry on the specified object.

ObjectAccessControlListMethodBuilder
Retrieves ACL entries on the specified object.

ObjectAccessControlMethodsBuilder
A builder providing access to all methods supported on objectAccessControl resources. It is not used directly, but through the Storage hub.

ObjectAccessControlPatchMethodBuilder
Updates an ACL entry on the specified object. This method supports patch semantics.

ObjectAccessControlProjectTeam
The project team associated with the entity, if any.

This type is not used in any activity, and only used as part of another schema.

ObjectAccessControlUpdateMethodBuilder
Updates an ACL entry on the specified object.

ObjectAccessControls
An access-control list.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list object access controls (response)

list default object access controls (response)

ObjectComposeMethodBuilder
Concatenates a list of existing objects into a new object in the same bucket.

ObjectCopyMethodBuilder
Copies an object to a specified location. Optionally overrides metadata.

ObjectDeleteMethodBuilder
Deletes an object and its metadata. Deletions are permanent if versioning is not enabled for the bucket, or if the generation parameter is used.

ObjectGetMethodBuilder
Retrieves an object or its metadata.

ObjectInsertMethodBuilder
Stores a new object and metadata.

ObjectListMethodBuilder
Retrieves a list of objects matching the criteria.

ObjectMethodsBuilder
A builder providing access to all methods supported on object resources. It is not used directly, but through the Storage hub.

ObjectOwner
The owner of the object. This will always be the uploader of the object.

This type is not used in any activity, and only used as part of another schema.

ObjectPatchMethodBuilder
Updates an object's metadata. This method supports patch semantics.

ObjectUpdateMethodBuilder
Updates an object's metadata.

ObjectWatchAllMethodBuilder
Watch for changes on all objects in a bucket.

Objects
A list of objects.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list objects (response)

Storage
Central instance to access all Storage related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.