Struct storage1::ObjectMethodsBuilder

pub struct ObjectMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on object resources. It is not used directly, but through the Storage hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate storage1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Storage::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `compose(...)`, `copy(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)`, `update(...)` and `watch_all(...)`
// to build up your call.
let rb = hub.objects();

Methods

impl<'a, C, NC, A> ObjectMethodsBuilder<'a, C, NC, A>

fn get(&self, bucket: &str, object: &str) -> ObjectGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves an object or its metadata.

fn watch_all(&self, request: &Channel, bucket: &str) -> ObjectWatchAllMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Watch for changes on all objects in a bucket.

fn update(&self, request: &Object, bucket: &str, object: &str) -> ObjectUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an object's metadata.

fn insert(&self, request: &Object, bucket: &str) -> ObjectInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Stores a new object and metadata.

fn compose(&self, request: &ComposeRequest, destination_bucket: &str, destination_object: &str) -> ObjectComposeMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Concatenates a list of existing objects into a new object in the same bucket.

fn delete(&self, bucket: &str, object: &str) -> ObjectDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes an object and its metadata. Deletions are permanent if versioning is not enabled for the bucket, or if the generation parameter is used.

fn list(&self, bucket: &str) -> ObjectListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves a list of objects matching the criteria.

fn copy(&self, request: &Object, source_bucket: &str, source_object: &str, destination_bucket: &str, destination_object: &str) -> ObjectCopyMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Copies an object to a specified location. Optionally overrides metadata.

fn patch(&self, request: &Object, bucket: &str, object: &str) -> ObjectPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an object's metadata. This method supports patch semantics.