Struct storage1::Storage

pub struct Storage<C, NC, A> {
    // some fields omitted
}

Central instance to access all Storage related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate storage1;
use storage1::ObjectAccessControl;
use storage1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use storage1::Storage;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Storage::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: ObjectAccessControl = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.object_access_controls().update(&req, "bucket", "object", "entity")
             .generation("erat")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Methods

impl<'a, C, NC, A> Storage<C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>>, A: GetToken

fn new(client: C, authenticator: A) -> Storage<C, NC, A>

fn bucket_access_controls(&'a self) -> BucketAccessControlMethodsBuilder<'a, C, NC, A>

fn buckets(&'a self) -> BucketMethodsBuilder<'a, C, NC, A>

fn channels(&'a self) -> ChannelMethodsBuilder<'a, C, NC, A>

fn default_object_access_controls(&'a self) -> DefaultObjectAccessControlMethodsBuilder<'a, C, NC, A>

fn object_access_controls(&'a self) -> ObjectAccessControlMethodsBuilder<'a, C, NC, A>

fn objects(&'a self) -> ObjectMethodsBuilder<'a, C, NC, A>